+++ src/platform/roblox/RobloxCompletion.cpp
@@ -49,9 +49,9 @@ static constexpr const char* COMMON_SERVICE_PROVIDER_PROPERTIES[] = {
     "GetService",
 };
 
-static lsp::CompletionItem createSuggestService(const std::string& service, size_t lineNumber, bool appendNewline = false)
+static lsp::CompletionItem createSuggestService(const std::string& service, size_t lineNumber, bool appendNewline = false, bool useConst = false)
 {
-    auto textEdit = Luau::LanguageServer::AutoImports::createServiceTextEdit(service, lineNumber, appendNewline);
+    auto textEdit = Luau::LanguageServer::AutoImports::createServiceTextEdit(service, lineNumber, appendNewline, useConst);
 
     lsp::CompletionItem item;
     item.label = service;
@@ -271,7 +271,7 @@ void RobloxPlatform::handleSuggestImports(const TextDocument& textDocument, cons
                 importsVisitor.firstRequireLine.value() - lineNumber == 0)
                 appendNewline = true;
 
-            items.emplace_back(createSuggestService(service, lineNumber, appendNewline));
+            items.emplace_back(createSuggestService(service, lineNumber, appendNewline, config.completion.imports.useConst));
         }
     }